% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Compute the adjacency matrix}
\usage{
adjacency_matrix(lst, fun = any_common, ...)
}
\arguments{
\item{lst}{A list of atomic vectors.}

\item{fun}{A predicate function of (at least) two vectors to be compared.}

\item{...}{Additional arguments passed on to the function \code{fn}.}
}
\value{
An adjacency matrix of boolean values.
}
\description{
Compute the adjacency matrix of a list of vectors.
}
