% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_common.R
\name{any_common}
\alias{any_common}
\title{Any element in common?}
\usage{
any_common(x, y, rm_na = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{y}{Another vector.}

\item{rm_na}{A logical indicating whether to remove \code{NA} elements before
checking for common elements between the two vectors.}
}
\value{
A logical scalar. \code{TRUE} indicates that \code{x} and \code{y}
  have at least one element in common, and \code{FALSE} otherwise.
}
\description{
Checks if two vectors have an element in common.
}
