% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{available_all}
\alias{available_all}
\alias{available_logicals}
\alias{available_models}
\alias{available_reporters_iso}
\alias{available_sections_code}
\alias{available_commodities_code}
\alias{available_vaccine}
\alias{available_yrs}
\alias{available_yrs_min}
\alias{available_yrs_max}
\title{Available UI parameters expressed as functions}
\usage{
available_all()

available_logicals()

available_models()

available_reporters_iso()

available_sections_code()

available_commodities_code()

available_vaccine()

available_yrs()

available_yrs_min()

available_yrs_max()
}
\description{
Available UI parameters expressed as functions
}
