% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{col_12}
\alias{col_12}
\alias{col_10}
\alias{col_9}
\alias{col_8}
\alias{col_6}
\alias{col_4}
\alias{col_3}
\alias{col_2}
\alias{col_1}
\title{Columns wrappers}
\usage{
col_12(...)

col_10(...)

col_9(...)

col_8(...)

col_6(...)

col_4(...)

col_3(...)

col_2(...)

col_1(...)
}
\arguments{
\item{...}{additional parameters for a column division}
}
\description{
Columns wrappers
}
