% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_by_common.R
\name{complete_by_common}
\alias{complete_by_common}
\title{Complete vectors by association of common elements}
\usage{
complete_by_common(lst, flatten = TRUE)
}
\arguments{
\item{lst}{A list of vectors to be compared.}

\item{flatten}{If \code{flatten} is \code{TRUE}, the returned list will have
the same depth as the input list \code{lst}. If \code{FALSE}, the returned
list retains contains one nesting level reflecting the grouping of vectors
with common elements.}
}
\value{
A list whose vectors have been expanded to include elements from
  other vectors.
}
\description{
This function completes vectors by adding elements from other vectors that
share at least one element in common. Note that is process is iterative,
where testing for common elements is redone as new elements are added.
}
\details{
This function takes a list \code{lst} of vectors. These vectors are compared
against each other and those pairs with common elements get are appended the
elements from the corresponding pair.
}
