% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{copy.attributes}
\alias{copy.attributes}
\title{Copy attributes between two data frames}
\usage{
copy.attributes(df1, df2)
}
\arguments{
\item{df1}{A data frame to copy attributes from.}

\item{df2}{A data frame to copy attributes to.}
}
\value{
The data frame in the \code{df2} parameter, with updated
attributes from \code{df1}.
}
\description{
A function to copy column attributes from one
data frame to another.  The function will copy all attributes attached
to each column.  The column order does not matter, and the data frames
do not need identical structures.
}
\examples{
# Prepare data
dat1 <- mtcars
dat2 <- mtcars

# Set labels for dat1
labels(dat1) <- list(mpg = "Miles Per Gallon",
                     cyl = "Cylinders",
                     disp = "Displacement")

# Copy labels from dat1 to dat2
dat2 <- copy.attributes(dat1, dat2)

# View results
labels(dat2)
# $mpg
}
