% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{custom_tooltip_short}
\alias{custom_tooltip_short}
\title{Custom Short Tooltip (For Highcharter Visuals)}
\usage{
custom_tooltip_short()
}
\description{
Custom Short Tooltip (For Highcharter Visuals)
}
