% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{enurl}
\alias{enurl}
\title{Create an url}
\usage{
enurl(url, text)
}
\arguments{
\item{url}{the URL}

\item{text}{the text to display}
}
\value{
an a tag
}
\description{
Create an url
}
\examples{
enurl("https://www.thinkr.fr", "ThinkR")
}
