% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Expand vectors that have been grouped together}
\usage{
expand(lst)
}
\arguments{
\item{lst}{A list whose elements are lists of grouped vectors. Each group of
vectors are expanded to include the elements of all the other vectors in
the same group.}
}
\value{
A list whose elements are lists of grouped vectors that have been
  expanded in a groupwise fashion.
}
\description{
Expand vectors that have been grouped together
}
