% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{gdp_deflator_adjustment}
\alias{gdp_deflator_adjustment}
\title{Convert dollars from year X to year Y}
\usage{
gdp_deflator_adjustment(d, reference_year, sql_con)
}
\arguments{
\item{d}{input dataset}

\item{reference_year}{year to convert dollars}

\item{sql_con}{SQL connection}
}
\description{
Convert dollars from year X to year Y
}
