% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{\%p\%}
\alias{\%p\%}
\title{An infix operator for paste0()}
\usage{
x \%p\% y
}
\arguments{
\item{x}{A value for the left side of the paste infix operator.}

\item{y}{A value for the right side of the paste infix operator.}
}
\value{
The concatenated or pasted value.
}
\description{
This function provides an infix operator for the
\code{\link{paste0}} function to concatenate strings.
}
