% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_elements.R
\name{group_elements}
\alias{group_elements}
\title{Determine the group elements}
\usage{
group_elements(lst, fun = any_common, ...)
}
\arguments{
\item{lst}{A list of atomic vectors.}

\item{fun}{A predicate function of (at least) two vectors to be compared.}

\item{...}{Additional arguments passed on to the function \code{fn}.}
}
\value{
A list of groups. Each element of the list is a group. A group is a
  list of vectors.
}
\description{
Determine the group elements.
}
