% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups_from_adjacency_matrix}
\alias{groups_from_adjacency_matrix}
\title{Determine groups from an adjacency matrix}
\usage{
groups_from_adjacency_matrix(m)
}
\arguments{
\item{m}{An adjacency matrix of \code{logical} type.}
}
\value{
A list of groups. Each element of the list is a vector of indices.
}
\description{
Determine groups from an adjacency matrix.
}
