\name{log.debug}
\alias{log.debug}
\title{Prints a debug message.}
\usage{
  log.debug(msg, ...)
}
\arguments{
  \item{msg}{the message that should be printed.}

  \item{...}{any additional text that should be printed.}
}
\description{
  This function will print a debug message.
}
\examples{
\dontrun{
log.debug("variable", 5)
}
}
\author{
  Rob Kooper
}

