\name{log.error}
\alias{log.error}
\title{Prints an error message.}
\usage{
  log.error(msg, ...)
}
\arguments{
  \item{msg}{the message that should be printed.}

  \item{...}{any additional text that should be printed.}
}
\description{
  This function will print an error message.
}
\examples{
\dontrun{
log.error("system did not converge")
}
}
\author{
  Rob Kooper
}

