% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{od_colors}
\alias{od_colors}
\title{Origin-Destination Colours (For Highcharter Visuals)}
\usage{
od_colors(d, con)
}
\arguments{
\item{d}{input dataset}

\item{con}{SQL connection}
}
\description{
Origin-Destination Colours (For Highcharter Visuals)
}
