% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{od_order_and_add_continent}
\alias{od_order_and_add_continent}
\title{Origin-Destination Order and Add Continent (For Highcharter Visuals)}
\usage{
od_order_and_add_continent(d, col = "trade_value_usd_exp", con)
}
\arguments{
\item{d}{input dataset}

\item{col}{column to collapse}

\item{con}{SQL connection}
}
\description{
Origin-Destination Order and Add Continent (For Highcharter Visuals)
}
