% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{od_to_highcharts}
\alias{od_to_highcharts}
\title{Origin-Destination to Highcharter (For Highcharter Visuals)}
\usage{
od_to_highcharts(d, d2)
}
\arguments{
\item{d}{input dataset for values}

\item{d2}{input dataset for colours}
}
\description{
Origin-Destination to Highcharter (For Highcharter Visuals)
}
