% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsshinycommon-package.R
\docType{package}
\name{otsshinycommon-package}
\alias{otsshinycommon}
\alias{otsshinycommon-package}
\title{otsshinycommon: Open Trade Statistics Helpers and Common Functions/Data}
\description{
Functions and data shared across the different Shiny apps that create tradestatistics.io.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tradestatistics/visualization-with-shiny}
  \item Report bugs at \url{https://github.com/tradestatistics/visualization-with-shiny/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mavargas11@uc.cl}

}
\keyword{internal}
