% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{p_aggregate_by_section}
\alias{p_aggregate_by_section}
\title{Add Percentages to Sections (For Highcharter Visuals)}
\usage{
p_aggregate_by_section(d, col, con)
}
\arguments{
\item{d}{input dataset}

\item{col}{column to collapse}

\item{con}{SQL connection}
}
\description{
Add Percentages to Sections (For Highcharter Visuals)
}
