% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{p_aggregate_products}
\alias{p_aggregate_products}
\title{Aggregate Products (For Highcharter Visuals)}
\usage{
p_aggregate_products(d, con)
}
\arguments{
\item{d}{input dataset}

\item{con}{SQL connection}
}
\description{
Aggregate Products (For Highcharter Visuals)
}
