% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{p_colors}
\alias{p_colors}
\title{Colorize Products (For Highcharter Visuals)}
\usage{
p_colors(d, con)
}
\arguments{
\item{d}{input dataset}

\item{con}{SQL connection}
}
\description{
Colorize Products (For Highcharter Visuals)
}
