% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{p_fix_section_and_aggregate}
\alias{p_fix_section_and_aggregate}
\title{Add Percentages to Sections (For Highcharter Visuals)}
\usage{
p_fix_section_and_aggregate(d, col, sql_con)
}
\arguments{
\item{d}{input dataset}

\item{col}{column to collapse}

\item{sql_con}{SQL connection}
}
\description{
Add Percentages to Sections (For Highcharter Visuals)
}
