% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{p_to_highcharts}
\alias{p_to_highcharts}
\title{Product to Highcharter (For Highcharter Visuals)}
\usage{
p_to_highcharts(d, d2)
}
\arguments{
\item{d}{input dataset for values}

\item{d2}{input dataset for colours}
}
\description{
Product to Highcharter (For Highcharter Visuals)
}
