% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{rv}
\alias{rv}
\alias{rvtl}
\title{Typing reactiveValues is too long}
\usage{
rv(...)

rvtl(...)
}
\arguments{
\item{...}{elements to pass to the function}
}
\description{
Typing reactiveValues is too long
}
