\name{read.settings.R}
\alias{read.settings.R}
\title{Generic function to open and parse PEcAn XML config/settings file

NEED MORE HERE. WORKING DRAFT OF SCRIPT TO OPEN PECAN SETTINGS}
\description{
  Generic function to open and parse PEcAn XML
  config/settings file

  NEED MORE HERE. WORKING DRAFT OF SCRIPT TO OPEN PECAN
  SETTINGS
}

