% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{roundup}
\alias{roundup}
\title{Rounds numbers up}
\usage{
roundup(x, digits = 0)
}
\arguments{
\item{x}{A vector of values to round.}

\item{digits}{A number of decimal places to round to. Default is zero.}
}
\value{
The rounded data vector.
}
\description{
A function to round numbers up where the last digit is
5 or above.  The function contains a parameter to control the number of
decimals to round to.
}
\examples{
# Round to even
round(2.4)  # 2
round(2.5)  # 2
round(2.6)  # 3

# Round up
roundup(2.4) # 2
roundup(2.5) # 3
roundup(2.6) # 3

}
