% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{sort.data.frame}
\alias{sort.data.frame}
\title{Orders a data frame}
\usage{
\method{sort}{data.frame}(
  x,
  decreasing = FALSE,
  ...,
  by = NULL,
  descending = FALSE,
  na.last = TRUE,
  index.return = FALSE
)
}
\arguments{
\item{x}{The input data to sort.}

\item{decreasing}{Whether the sort should be in decreasing order.}

\item{...}{Follow on parameters.}

\item{by}{A vector of column names to sort by.  If this parameter
is not supplied, the function will sort by all columns in order
from left to right.}

\item{descending}{A vector of TRUE or FALSE values corresponding
to the variables in the \code{by} parameter.  These values will determine
the direction to sort each column.}

\item{na.last}{Whether to put NA values last in the sort.}

\item{index.return}{Whether to return the sorted data frame or an index}
}
\value{
A vector of row numbers in the desired sort order.
}
\description{
An override to the Base R order function that actually works
with data frames.  Allows multiple columns to be sorted, and multiple
descending options.
}
