% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{spaces}
\alias{spaces}
\title{Creates a string of blank spaces}
\usage{
spaces(num)
}
\arguments{
\item{num}{The desired number of spaces.}
}
\value{
A single character vector of blank spaces.
}
\description{
A function to create a string of some number of blank spaces.
This function is useful when trying to align things.
}
\examples{
# Create spaces
spaces(10)
# [1] "          "

# Use spaces to separate something
str <- "Left" \%p\% spaces(40) \%p\% "Right"
str
# [1] "Left                                        Right"
}
\concept{strings}
