% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syspath.R
\name{Sys.path.experimental}
\alias{Sys.path.experimental}
\title{EXPERIMENTAL: Returns the path of the current program}
\usage{
Sys.path.experimental()
}
\value{
The full path of the currently running program, or a NULL.
}
\description{
A function that gets the full path of the currently running
program.  If the function fails to retrieve the path for some reason,
it will return a NULL.  The function takes no parameters.
}
\examples{
# Get current path
pth <- Sys.path.experimental()
pth
# [1] "C:/programs/myprogram.R"
}
\seealso{
Other fileops: 
\code{\link{Sys.path}()},
\code{\link{dir.find}()},
\code{\link{file.find}()}
}
\concept{fileops}
