% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{Sys.path}
\alias{Sys.path}
\title{Returns the path of the current program}
\usage{
Sys.path()
}
\value{
The full path of the currently running program, or a NULL.
}
\description{
A function that gets the full path of the currently running
program.  If the function fails to retrieve the path for some reason,
it will return a NULL.  The function takes no parameters.

Credit for this function goes to Andrew Simmons and the
\code{\link{this.path}} package.
}
\examples{
# Get current path
pth <- Sys.path()
pth
# [1] "C:/programs/myprogram.R"
}
