% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{v}
\alias{v}
\title{Combine unquoted values}
\usage{
v(...)
}
\arguments{
\item{...}{One or more unquoted values.}
}
\value{
A vector of quoted values.
}
\description{
A function to quote and combine unquoted values.
The function will return a vector of quoted values.  This function
allows you to use non-standard evaluation for any parameter
that accepts a string or vector of strings.
}
\examples{
# Combine unquoted values
v(var1, var2, var3)
# [1] "var1" "var2" "var3"
}
