% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{p_to_highcharts}
\alias{p_to_highcharts}
\title{Product to Highcharter (For Highcharter Visuals)}
\usage{
p_to_highcharts(d, d2 = NULL)
}
\arguments{
\item{d}{input dataset for values}

\item{d2}{optional dataset for colours (when provided it will be used as the
color reference so the same `section_name` maps to the same colour across
different treemaps)}
}
\description{
Product to Highcharter (For Highcharter Visuals)
}
