% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{roundup}
\alias{roundup}
\title{Rounds numbers up}
\usage{
roundup(x, digits = 0)
}
\arguments{
\item{x}{A vector of values to round.}

\item{digits}{A number of decimal places to round to. Default is zero.}
}
\value{
The rounded data vector.
}
\description{
A function that rounds positive numbers up when the last digit
is a 5.  For negative numbers ending in 5, the function actually rounds down.
"Round away from zero" is the most accurate description of this function.
}
\examples{
# Round to even
round(2.4)   # 2
round(2.5)   # 2
round(-2.5)  # -2
round(2.6)   # 3

# Round up
roundup(2.4)  # 2
roundup(2.5)  # 3
roundup(-2.5) # -3
roundup(2.6)  # 3

}
